bugfix after refactoring, remove validation on system field to allow save empty value. ViewModel config instead of Block
