+++ src/etc/adminhtml/system.xml
@@ -28,7 +28,6 @@
                     <label>List of Events</label>
                     <config_path>perspective/settings/forwarding_events_list</config_path>
                     <source_model>Perspective\Partytown\Model\Adminhtml\System\Config\ForwardingMultiselect</source_model>
-                    <validate>validate-select</validate>
                     <can_be_empty>1</can_be_empty>
                     <comment>A list of events representing function calls on the main thread to forward to the web worker</comment>
                 </field>
@@ -62,7 +61,6 @@
                     <label>Log configuration properties</label>
                     <config_path>perspective/settings/debug_logging_list</config_path>
                     <source_model>Perspective\Partytown\Model\Adminhtml\System\Config\LogConfigMultiselect</source_model>
-                    <validate>validate-select</validate>
                     <can_be_empty>1</can_be_empty>
                     <comment><![CDATA[If you are using Chrome DevTools to see these logs, make sure that the Verbose log level is selected in Console. <br>  <a target="_blank" href="https://partytown.builder.io/debugging" title="Read more">Read more</a>]]></comment>
                 </field>
