+++ src/view/frontend/layout/default.xml
@@ -8,10 +8,13 @@
                        ifconfig="perspective/settings/status"
             >
                 <block name="perspective.partytown.module.init"
-                   class="Perspective\Partytown\Block\Config" 
-                   template="Perspective_Partytown::js/partytown/jsConfig.phtml"
-                   as="js-config"
-                />
+                       template="Perspective_Partytown::js/partytown/jsConfig.phtml"
+                       as="js-config"
+                >
+                    <arguments>
+                        <argument name="view_model" xsi:type="object">Perspective\Partytown\ViewModel\PartytownConfig</argument>
+                    </arguments>
+                </block>
             </block>
         </referenceContainer>
     </body>
